+++ models/CustomPage.php
@@ -92,12 +92,17 @@ class CustomPage extends ContentActiveRecord implements ViewableInterface, Edita
     /**
      * @var array Groups for custom visibility restriction
      */
-    public $visibility_groups;
+    public $visibilityGroups;
 
     /**
      * @var array Languages for custom visibility restriction
      */
-    public $visibility_languages;
+    public $visibilityLanguages;
+
+    /**
+     * @var bool True when the page is visible only for mobile app users
+     */
+    public $visibilityMobileApp;
 
     private ?VisibilityService $_visibilityService = null;
     private ?SettingService $_settingService = null;
@@ -156,8 +161,9 @@ class CustomPage extends ContentActiveRecord implements ViewableInterface, Edita
             'targetUrl' => Yii::t('CustomPagesModule.base', 'Target Url'),
             'templateId' => Yii::t('CustomPagesModule.base', 'Template Layout'),
             'visibility' => Yii::t('CustomPagesModule.model', 'Visibility'),
-            'visibility_groups' => Yii::t('CustomPagesModule.model', 'Visible to Group Members'),
-            'visibility_languages' => Yii::t('CustomPagesModule.model', 'Language-Based Visibility'),
+            'visibilityGroups' => Yii::t('CustomPagesModule.model', 'Visible to Group Members'),
+            'visibilityLanguages' => Yii::t('CustomPagesModule.model', 'Language-Based Visibility'),
+            'visibilityMobileApp' => Yii::t('CustomPagesModule.model', 'Mobile App Users'),
             'editors' => Yii::t('CustomPagesModule.model', 'Editors'),
             'hide_menu' => Yii::t('CustomPagesModule.model', 'Hide in Navigation'),
         ];
@@ -196,7 +202,8 @@ class CustomPage extends ContentActiveRecord implements ViewableInterface, Edita
             [['target'], 'validateTarget'],
             [['type'], 'validateContentType'],
             [['visibility'], 'integer', 'min' => self::VISIBILITY_PRIVATE, 'max' => self::VISIBILITY_CUSTOM],
-            [['visibility_groups', 'visibility_languages', 'editors'], 'safe'],
+            [['visibilityGroups', 'visibilityLanguages', 'editors'], 'safe'],
+            [['visibilityMobileApp'], 'boolean'],
             [['title', 'target'], 'string', 'max' => 255],
         ];
 
@@ -306,7 +313,7 @@ class CustomPage extends ContentActiveRecord implements ViewableInterface, Edita
     {
         parent::afterFind();
 
-        $this->visibilityService->loadAdditionalOptions();
+        $this->settingService->initAll();
     }
 
     /**
@@ -334,7 +341,7 @@ class CustomPage extends ContentActiveRecord implements ViewableInterface, Edita
             RichText::postProcess($this->abstract, $this);
         }
 
-        $this->visibilityService->updateAdditionalOptions();
+        $this->settingService->updateAll();
     }
 
     /**